test(a11y): consolidate app component accessibility audits into a11y.spec
