+++ test/nuxt/components/app/HeaderAuth.spec.ts
@@ -1,7 +1,6 @@
 import { AppHeaderAuth } from "#components";
 import { mountSuspended } from "@nuxt/test-utils/runtime";
 import { describe, expect, it } from "vitest";
-import { runAxe } from "../../utils/axe";
 
 // `AppHeaderAuth` is a `.client.vue` component that reads the session through the
 // auth module's `useAuth` -> `useUserSession`. Those calls resolve inside the
@@ -9,6 +8,7 @@ import { runAxe } from "../../utils/axe";
 // without a real session the component renders its logged-out state, which is
 // exactly what the server emits and what the client shows before the session
 // resolves. These tests cover that default state.
+// Accessibility audits for AppHeaderAuth live in test/nuxt/a11y.spec.ts.
 describe("AppHeaderAuth", () => {
 	describe("hydration safety", () => {
 		// Client-only component: never server-rendered, so a hydration mismatch is
@@ -19,23 +19,6 @@ describe("AppHeaderAuth", () => {
 			const first = await mountSuspended(AppHeaderAuth);
 			const second = await mountSuspended(AppHeaderAuth);
 			expect(second.html()).toBe(first.html());
-		});
-	});
-
-	describe("accessibility", () => {
-		it("has no axe-core violations", async () => {
-			const wrapper = await mountSuspended(AppHeaderAuth);
-			const results = await runAxe(wrapper);
-			expect(results.violations).toEqual([]);
-		});
-
-		it("renders sign-in buttons that link to /login with an accessible name", async () => {
-			const wrapper = await mountSuspended(AppHeaderAuth);
-			const signInLinks = wrapper.findAll("a[href='/login']");
-			expect(signInLinks.length).toBeGreaterThan(0);
-			for (const link of signInLinks) {
-				expect(link.attributes("aria-label")).toBe("Sign in with Discord");
-			}
 		});
 	});
 });
